RNIC-W02: create HomeScreen component with API call
